optimize render, data race fix for render, add auto add remove register entity for systems (only object in scene can be rendered or be clickable)

+++ src/constructorWindow.h
@@ -62,7 +62,7 @@ private:
     
     int lastFps = 0;
     int currentFps = 0;
-    int lastTimeCheckFps = 0;
+    long long lastTimeCheckFps = 0;
     
     std::atomic_int mouseX = 0;
     std::atomic_int mouseY = 0;
